refactor access request attachment button
